A template script
This script just has two states dead and alive denoted by 0 and 1 which is of datatype 'char' so I have converted the incoming message to 'str' since string manipulation and writing it to a file is lot easier. Hope you can build your logic on top of it or modify the code to implement your ideas.