Create Duel_simulation.py
The code is similar to covid 19 simulation where the character codes indicate what state the microbit is currently in. I have also included a probability engine not to kill the microbits at once and also to progress the game. You can flash this code to your microbits and experiment with some time intervals and change your simulation time. You can also change the logic as you like or according to your rules